webhook/requirements.txt
azure-functions>=1.17.0
pyodbc>=5.0.0
